requirement.txt
flask-jwt-extended
PyJWT==2.6.0
bcrypt==4.0.1
Flask==2.2.3
python-dotenv
pymongo
